share window on footer data page
