+++ js/src/templates/indicators/indicator.hbs
@@ -7,6 +7,15 @@
 <div class="l-toolbar -footer">
   <div class="wrap">
     <div class="js--legend"></div>
+
+    <div class="c-share">
+      <button class="js--view-share btn">
+        <svg class="icon -regular">
+          <use xmlns:xlink="http://www.w3.org/1999/xlink" xlink:href="#icon-share"></use>
+        </svg>
+        <span class="button-literal">share</span>
+      </button>
+    </div>
     {{!-- <div class="js--toolbar-utils"></div> --}}
   </div>
 </div>
